+++ packages/lwc-engine/src/framework/decorators/track.ts
@@ -3,7 +3,7 @@ import { isArray, isObject, defineProperty, isUndefined } from "../language";
 import { getReactiveProxy, isObservable } from "../reactive";
 import { isRendering, vmBeingRendered } from "../invoker";
 import { observeMutation, notifyMutation } from "../watcher";
-import { VMElement, VM } from "../vm";
+import { VMElement } from "../vm";
 import { getCustomElementVM } from "../html-element";
 
 // stub function to prevent misuse of the @track decorator