+++ packages/lwc-integration/src/components/rendering/test-toggle-empty-classname/toggle-empty-classname.spec.js
@@ -1,0 +1,17 @@
+const assert = require('assert');
+describe('Toggling to empty classname', () => {
+    const URL = 'http://localhost:4567/toggle-empty-classname';
+    let element;
+
+    before(() => {
+        browser.url(URL);
+    });
+
+    it('should have the right value', function () {
+        const element = browser.element('toggle-empty-classname');
+        element.click();
+
+        const className = browser.getAttribute('toggle-empty-classname div', 'class');
+        assert.deepEqual(className, '');
+    });
+});